% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_table.R
\name{gen_table}
\alias{gen_table}
\title{Generate a Table}
\usage{
gen_table(x)
}
\arguments{
\item{Input}{path}
}
\value{
A matrix of the infile
}
\description{
This function generates a multiplication table. It takes a user input
and auto generates table of the input number.
}
\keyword{table}
